# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GeminiChatMessageRole = typeof GeminiChatMessageRole[keyof typeof GeminiChatMessageRole];


export const GeminiChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface GeminiChatMessage {
  role: GeminiChatMessageRole;
  content: string;
}

export interface GeminiChatInput {
  messages: GeminiChatMessage[];
  /** Optional Free Practice topic (e.g. "Hometown", "Technology"). When present, the examiner must keep every question and follow-up strictly scoped to this topic instead of open-ended conversation. */
  topic?: string;
}

/**
 * Estimated IELTS Speaking band scores (0-9) for the student's last utterance.
 */
export interface BandScoreBreakdown {
  /** Fluency and Coherence band score (0-9). */
  fluency: number;
  /** Lexical Resource band score (0-9). */
  lexicalResource: number;
  /** Grammatical Range and Accuracy band score (0-9). */
  grammaticalRange: number;
  /** Pronunciation band score (0-9). */
  pronunciation: number;
  /** Overall band score (0-9), the average of the four criteria. */
  overall: number;
}

export interface VocabularyUpgrade {
  /** The simple word or phrase the student used. */
  original: string;
  /** A Band 7+/8+ synonym or idiomatic expression to replace it. */
  upgrade: string;
}

export interface GeminiChatOutput {
  /** The examiner's short conversational reply (speak this aloud). */
  reply: string;
  /** Correction of any grammar/word-choice mistakes in the student's last message, or confirmation it was correct. Display only, do not speak aloud. */
  correction: string;
  /** A more natural, idiomatic, or higher band-score way to phrase what the student said. Display only, do not speak aloud. */
  bandUpgrade: string;
  bandScores: BandScoreBreakdown;
  /** 1-2 simple words the student used, each paired with a Band 7+/8+ synonym or idiomatic replacement. */
  vocabularyUpgrades: VocabularyUpgrade[];
}

export interface GeminiError {
  error: string;
}