# lib/api-zod/src/generated/types/bandScoreBreakdown.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Estimated IELTS Speaking band scores (0-9) for the student's last utterance.
 */
export interface BandScoreBreakdown {
  /** Fluency and Coherence band score (0-9). */
  fluency: number;
  /** Lexical Resource band score (0-9). */
  lexicalResource: number;
  /** Grammatical Range and Accuracy band score (0-9). */
  grammaticalRange: number;
  /** Pronunciation band score (0-9). */
  pronunciation: number;
  /** Overall band score (0-9), the average of the four criteria. */
  overall: number;
}