# lib/api-zod/src/generated/types/geminiChatInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeminiChatMessage } from './geminiChatMessage';

export interface GeminiChatInput {
  messages: GeminiChatMessage[];
  /** Optional Free Practice topic (e.g. "Hometown", "Technology"). When present, the examiner must keep every question and follow-up strictly scoped to this topic instead of open-ended conversation. */
  topic?: string;
}