# lib/api-zod/src/generated/types/geminiChatOutput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BandScoreBreakdown } from './bandScoreBreakdown';
import type { VocabularyUpgrade } from './vocabularyUpgrade';

export interface GeminiChatOutput {
  /** The examiner's short conversational reply (speak this aloud). */
  reply: string;
  /** Correction of any grammar/word-choice mistakes in the student's last message, or confirmation it was correct. Display only, do not speak aloud. */
  correction: string;
  /** A more natural, idiomatic, or higher band-score way to phrase what the student said. Display only, do not speak aloud. */
  bandUpgrade: string;
  bandScores: BandScoreBreakdown;
  /** 1-2 simple words the student used, each paired with a Band 7+/8+ synonym or idiomatic replacement. */
  vocabularyUpgrades: VocabularyUpgrade[];
}