# lib/api-zod/src/generated/types/vocabularyUpgrade.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VocabularyUpgrade {
  /** The simple word or phrase the student used. */
  original: string;
  /** A Band 7+/8+ synonym or idiomatic expression to replace it. */
  upgrade: string;
}